Séparation de useItinerary en draw et fetch
