+++ app/voyage/itinerary/useDrawItinerary.ts
@@ -1,27 +1,26 @@
 import useSetSearchParams from '@/components/useSetSearchParams'
-import distance from '@turf/distance'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
+import { useMediaQuery } from 'usehooks-ts'
+import { buildAllezPart, removeStatePart } from '../SetDestination'
 import useDrawTransit from '../effects/useDrawTransit'
 import { decodeDate, initialDate } from './DateSelector'
-import { computeMotisTrip } from '@/app/voyage/itinerary/transit/motisRequest'
+import { letterFromIndex } from './Steps'
 import useDrawRoute from './useDrawRoute'
 import useFetchDrawBikeParkings from './useFetchDrawBikeParkings'
 import { geoSerializeSteps } from './areStepsEqual'
-import { buildAllezPart, removeStatePart } from '../SetDestination'
-import { letterFromIndex } from './Steps'
-import { useMediaQuery } from 'usehooks-ts'
 import { modeKeyFromQuery } from './Itinerary'
 
 export default function useItinerary(
 	map,
 	itineraryMode,
-	bikeRouteProfile,
 	searchParams,
 	state,
-	zoom
+	zoom,
+	routes
 ) {
 	const mode = modeKeyFromQuery(searchParams.mode)
 	const desktop = useMediaQuery('(min-width: 800px)')
+
 	useEffect(() => {
 		if (!map) return
 		if (state.length === 2 && state[0] == null && state[1] !== null) {
@@ -30,8 +29,6 @@ export default function useItinerary(
 		}
 	}, [state, map, desktop])
 	//TODO
-	const [routes, setRoutes] = useState(null)
-	const date = decodeDate(searchParams.date) || initialDate()
 	const selectedConnection = searchParams.choix
 
 	//const [motisTrips, setMotisTrips] = useState(null)
@@ -48,44 +45,9 @@ export default function useItinerary(
 
 	useFetchDrawBikeParkings(map, routes?.cycling)
 
-	const updateRoute = (key, value) =>
-		setRoutes((routes) => ({ ...(routes || {}), [key]: value }))
 	const setSearchParams = useSetSearchParams()
 
-	/*
-	 * {
-  "type": "Feature",
-  "geometry": {
-    "type": "Point",
-    "coordinates": [
-      -1.6369411434976087,
-      48.17932437715612
-    ]
-  },
-  "properties": {
-    "id": "1704742857502"
-  }
-}
-	 */
-	const serializedPoints = geoSerializeSteps(state)
-
-	const points = useMemo(() => {
-		const points = state
-			.map((step, index) => {
-				if (step == null) return
-				const { longitude, latitude, key } = step
-				return {
-					type: 'Feature',
-					geometry: {
-						type: 'Point',
-						coordinates: [+longitude, +latitude],
-					},
-					properties: { key, letter: letterFromIndex(index) },
-				}
-			})
-			.filter(Boolean)
-		return points
-	}, [serializedPoints])
+	const [serializedPoints, points] = useMemoPointsFromState(state)
 
 	const linestrings = useMemo(
 		() =>
@@ -199,106 +161,6 @@ export default function useItinerary(
 		}
 	}, [map, serializedPoints, setSearchParams, itineraryMode, oldAllez])
 
-	/* Routing requests are made here */
-	useEffect(() => {
-		if (points.length < 2) {
-			setRoutes(null)
-			return
-		}
-
-		async function fetchBrouterRoute(
-			points,
-			itineraryDistance,
-			profile,
-			maxDistance
-		) {
-			if (itineraryDistance > maxDistance) return null
-
-			const lonLats = points
-				.map(
-					({
-						geometry: {
-							coordinates: [lon, lat],
-						},
-					}) => `${lon},${lat}`
-				)
-				.join('|')
-			const url = `https://brouter.osc-fr1.scalingo.io/brouter?lonlats=${lonLats}&profile=${profile}&alternativeidx=0&format=geojson`
-			const res = await fetch(url)
-			const json = await res.json()
-			if (!json.features) return
-			return json
-		}
-
-		//TODO fails is 3rd point is closer to 1st than 2nd, use reduce that sums
-		const itineraryDistance = distance(points[0], points.slice(-1)[0])
-
-		const fetchRoutes = async () => {
-			updateRoute('cycling', 'loading')
-			const cycling = await fetchBrouterRoute(
-				points,
-				itineraryDistance,
-				bikeRouteProfile,
-				35 // ~ 25 km/h (ebike) x 1:30 hours
-			)
-			updateRoute('cycling', cycling)
-
-			updateRoute('walking', 'loading')
-			const walking = await fetchBrouterRoute(
-				points,
-				itineraryDistance,
-				'hiking-mountain',
-				mode === 'walking' ? Infinity : 4 // ~ 3 km/h donc 4 km = 1h20 minutes, au-dessus ça me semble peu pertinent de proposer la marche par défaut
-			)
-			updateRoute('walking', walking)
-		}
-		fetchRoutes()
-	}, [points, setRoutes, bikeRouteProfile, mode])
-
-	useEffect(() => {
-		if (points.length < 2) {
-			setRoutes(null)
-			return
-		}
-
-		async function fetchTrainRoute(multiplePoints, itineraryDistance, date) {
-			const minTransitDistance = 0.5 // please walk or bike
-			if (itineraryDistance < minTransitDistance) return null
-			const points =
-				multiplePoints.length > 2
-					? [multiplePoints[0], multiplePoints.slice(-1)[0]]
-					: multiplePoints
-			const lonLats = points.map(
-				({
-					geometry: {
-						coordinates: [lng, lat],
-					},
-				}) => ({ lat, lng })
-			)
-
-			const json = await computeMotisTrip(lonLats[0], lonLats[1], date)
-
-			if (json.state === 'error') return json
-
-			if (!json?.content) return null
-			/*
-			return sections.map((el) => ({
-				type: 'Feature',
-				properties: el.geojson.properties[0],
-				geometry: { coordinates: el.geojson.coordinates, type: 'LineString' },
-			}))
-			*/
-			return json.content
-		}
-		//TODO fails is 3rd point is closer to 1st than 2nd, use reduce that sums
-		const itineraryDistance = distance(points[0], points.slice(-1)[0])
-
-		updateRoute('transit', { state: 'loading' })
-		fetchTrainRoute(points, itineraryDistance, date).then((transit) =>
-			setRoutes((routes) => ({ ...routes, transit }))
-		)
-	}, [points, setRoutes, date])
-
 	// GeoJSON object to hold our measurement features
 
 	useEffect(() => {
@@ -321,6 +183,41 @@ export default function useItinerary(
 		: Math.round(rawDistance) + ' km'
 
 */
-	const resetItinerary = () => setSearchParams({ allez: undefined })
-	return [resetItinerary, routes, date]
+}
+
+export const useMemoPointsFromState = (state) => {
+	/*
+	 * {
+  "type": "Feature",
+  "geometry": {
+    "type": "Point",
+    "coordinates": [
+      -1.6369411434976087,
+      48.17932437715612
+    ]
+  },
+  "properties": {
+    "id": "1704742857502"
+  }
+}
+	 */
+	const serializedPoints = geoSerializeSteps(state)
+	const points = useMemo(() => {
+		const points = state
+			.map((step, index) => {
+				if (step == null) return
+				const { longitude, latitude, key } = step
+				return {
+					type: 'Feature',
+					geometry: {
+						type: 'Point',
+						coordinates: [+longitude, +latitude],
+					},
+					properties: { key, letter: letterFromIndex(index) },
+				}
+			})
+			.filter(Boolean)
+		return points
+	}, [serializedPoints])
+	return [serializedPoints, points]
 }
